package_requirements.txt
attrs
pandas>=0.23
scikit-learn
wrapt
variants
